Improve Queries and Dashboards testing
