+++ lib/sanbase/queries/queries.ex
@@ -125,9 +125,9 @@ defmodule Sanbase.Queries do
   Get a query in order to read or run it.
   This can be done by owner or by anyone if the query is public.
   """
-  @spec get_query(query_id, user_id) :: {:ok, Query.t()} | {:error, String.t()}
-  def get_query(query_id, querying_user_id) do
-    query = Query.get_for_read(query_id, querying_user_id)
+  @spec get_query(query_id, user_id, Keyword.t()) :: {:ok, Query.t()} | {:error, String.t()}
+  def get_query(query_id, querying_user_id, opts \\ []) do
+    query = Query.get_for_read(query_id, querying_user_id, opts)
 
     case Repo.one(query) do
       %Query{} = query -> {:ok, query}
@@ -381,22 +381,51 @@ defmodule Sanbase.Queries do
   end
 
   @doc ~s"""
-  Cache a query execution
-  TODO: Add more documentation
+  Cache a query execution.
+  A query can be cached by its owner, as well as any other user who sees the query.
+  The query cache is stored under the (query_id, user_id) complex key, so the check is
+  only if the user can *read* the query, not mutate it.
   """
   @spec cache_query_execution(query_id, any(), user_id, Keyword.t()) :: any()
-  def cache_query_execution(
-        query_id,
-        query_result,
-        user_id,
-        opts \\ []
-      ) do
-    Queries.Cache.create_or_update_cache(
-      query_id,
-      query_result,
-      user_id,
-      opts
-    )
+  def cache_query_execution(query_id, query_result, user_id, opts \\ []) do
+    Ecto.Multi.new()
+    |> Ecto.Multi.run(:get_for_read, fn _, _ ->
+      # This serves the purpose of checking if the querying user has
+      # read access to the query. Only read access is needed to create
+      # a cache, as the cache is not a part of the query itself and is
+      # stored under the (query_id, user_id) complex key.
+      get_query(query_id, user_id, preload?: false)
+    end)
+    |> Ecto.Multi.run(:create_or_update_cache, fn _, %{get_for_read: query} ->
+      # This also executes a transaction inside it and it is ok as
+      # ecto transactions can be nested
+      Queries.Cache.create_or_update_cache(query, query_result, user_id, opts)
+    end)
+    |> Sanbase.Repo.transaction()
+    |> process_transaction_result(:create_or_update_cache)
+  end
+
+  def get_cached_query_executions(query_id, user_id) do
+    Ecto.Multi.new()
+    |> Ecto.Multi.run(:get_for_read, fn _, _ ->
+      get_query(query_id, user_id, preload?: false)
+    end)
+    |> Ecto.Multi.run(:get_cached_executions, fn _, %{get_for_read: query} ->
+      Queries.Cache.get_cached_executions(query, user_id)
+    end)
+    |> Ecto.Multi.run(:add_hash_match_field, fn _, changes ->
+      %{get_for_read: query, get_cached_executions: cached_executions} = changes
+      current_query_hash = Query.hash(query)
+
+      result =
+        Enum.map(cached_executions, fn execution ->
+          Map.put(execution, :is_query_hash_matching, execution.query_hash == current_query_hash)
+        end)
+
+      {:ok, result}
+    end)
+    |> Sanbase.Repo.transaction()
+    |> process_transaction_result(:add_hash_match_field)
   end
 
   # Private functions
