+++ test/sanbase_web/graphql/comments/short_url_comments_api_test.exs
@@ -4,7 +4,7 @@ defmodule SanbaseWeb.GraphqlShortUrlCommentsApiTest do
   import Sanbase.Factory
   import Sanbase.TestHelpers
   import SanbaseWeb.Graphql.TestHelpers
-  import Sanbase.CommentsApiHelper
+  import SanbaseWeb.CommentsApiHelper
 
   @opts [entity_type: :short_url, extra_fields: ["shortUrlId"]]
   setup do
