+++ test/sanbase_web/graphql/comments/timeline_event_comments_api_test.exs
@@ -4,7 +4,7 @@ defmodule SanbaseWeb.GraphqlTimelineEventCommentsApiTest do
   import Sanbase.Factory
   import Sanbase.TestHelpers
   import SanbaseWeb.Graphql.TestHelpers
-  import Sanbase.CommentsApiHelper
+  import SanbaseWeb.CommentsApiHelper
 
   @opts [entity_type: :timeline_event, extra_fields: ["timelineEventId"]]
   setup do
